Fixes to cleanData.py
MVP data and Season Stats do not line up, since MVP data uses the former year of the season and season stats uses the latter year of the season. This change fixes this issue by adding 1 to the MVP data years, and removes the mvp data for the 2017-2018 season (since season stats does not include this).